Re-Add BayTrail Soc GPIO and LPSS ACPI IDs.
The ACPI ID of gpio driver was removed because there was a bug triggered when gpio driver registered handlers for the ACPI signalled gpio interrupts (caused graphics to crash).
The bug is reported here:
https://bugzilla.kernel.org/show_bug.cgi?id=68291
This patch is just a workaround to add the ACPI IDs back to let the gpio work on BayTrail.

Change-Id: Id27427eb6f00a664db4de4ba1d2a71df284f7b93