FROM_UPSTREAM [VPG]: drm: Rename DP training vswing pre emph
Adding new defines, older one will be removed in the last patch in the series.
This is to rename the defines to have levels instead of values for vswing and
pre-emph levels as the values may differ in other scenarios like low vswing of
eDP1.4 where the values are different.

Done using following cocci patch for each define:
@@
@@

 # define DP_TRAIN_VOLTAGE_SWING_400     (0 << 0)
+ # define DP_TRAIN_VOLTAGE_SWING_LEVEL_0     (0 << 0)

...
Cc: dri-devel@lists.freedesktop.org

Upstream-repo: git://anongit.freedesktop.org/drm-intel
Upstream-commit: 0504cd170c86f0617f27a31f1ba2328caa976707
Upstream-as-of-tag: drm-intel-next-2014-08-22-11456-g0504cd1
Change-Id: Ibec9ab04e054e4f3bb0a8db23d8c4a1d28f6b7c5
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5971
Signed-off-by: Sonika Jindal <sonika.jindal@intel.com>
Acked-by: Dave Airlie <airlied@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>