FOR_UPSTREAM [VPG]: drm/i915: capture more detail in ringbuf & LRC request error dump
When analysing error dumps, it's useful to know the state of the
ringbuf structures, such as last_retired_head. Also, in LRC mode,
we need to know the context as well as the tail pointer for each
request in the request queue. This code captures and dumps the
extra information.

v2: (Tomas Elf)
Fixed compiler warning in err_printf format string (%x -> %p).
Fixed various type inconsistencies.

Signed-off-by: Dave Gordon <david.s.gordon@intel.com>
Signed-off-by: Tomas Elf <tomas.elf@intel.com>
Issue: GMIN-3575
Tracked-On: https://jira01.devtools.intel.com/browse/GMIN-3575
Change-Id: Iaf92fd081f2a7ba99eee7284c1f601e221249b1f