Black screen when recording dark environment
When we start to record on front camera, we get a black screen
during all the record if the environment was too dark when we
started it. When exposure is reaching a threshold and at the same
time we start a record, then the driver is not able to refresh
exposure value.
This patch aims to fix this issue, registering the value of the
first exposure before starting the record and then resetting it
with the same value once the record has been started.

(Forward-port to recent ICG changes by Andy Ross)

Issue: GMIN-3468
Change-Id: I8da2105154808e143a035a1dcbcba677aede78de
Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>