usb: dwc3: ugly hack: implement watchdog
We have an stability issue going on with dwc3 driver. We don't know the exact
cause yet, but we isolated the problem. It happens because either the
controller or phy is getting stuck just after we receive USB RESET event:
instead of the connection start right away, we don't see any usb activity. The
only solution so far is to do a soft reset.

This watchdog is enabled as soon as we receive the USB RESET event. If no
connecting comes in 1 second, the watchdog power cycles the controller.

Change-Id: I5da255e978e21c64dd05aaad2eb79cf9f87d7e38
Signed-off-by: David A Cohen <david.a.cohen@intel.com>
Signed-off-by: Fei Yang <fei.yang@intel.com>
Signed-off-by: Eric Ernst <eric.ernst@intel.com>