MUST_REBASE [VPG]: video/adf/intel: clean up MIPI DSI pipe vs port usage
This patch is in preparation of DSI dual link feature. MIPI DSI works on
Ports A and C, which in hardware maps to Pipe A and Pipe B respectively.
With introduction of DSI dual link, things will get more complicated,
so cleaned up the register definitions and the code.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed

For: GMINL-5276
Change-Id: Ib1d569c3f90145fea0745f3ec4fbead510d0685c
Signed-off-by: Gaurav K Singh <gaurav.k.singh@intel.com>