SQUASHME! [VPG]: drm/i915: Keep Ring HWS page address persistent across GPU reset
After an Engine is reset it is found that HW is losing Hardware Status Page address
loaded in the register. We store various values in HWS, one of it is the
active seqno. Before a batch is executed it's seqno is written at an index
in this page and it is cleared after it is done.

In the case of a hang, the batch that caused the hang would not be able
to clear this field. Native sync logic uses this active seqno and signals
the batch as timed out as it couldn't complete its execution properly.
Since HW was losing its address it was not able to signal its status correctly.
The fix is to save its value along with head, tail and control register
and restore it after reset.

Noticed a duplicate variable in Ring hangcheck structure to track
last known head position. The changes to remove it are added to this
patch as both of them are getting squashed into the same patch.

This patch should be squashed into the following patch:

     Author: Tomas Elf <tomas.elf@intel.com>
     Date:   Fri Oct 24 13:57:09 2014 +0100
     Change-Id: I45de965b4a81d93c135a74248eb22a8a5c816e9f
     FOR_UPSTREAM [VPG]: drm/i915: Added gen8 support for Timeout Detection Recovery for engine hangs

For: GMIN-3564
Tracked-On: https://jira01.devtools.intel.com/browse/GMIN-3564
Signed-off-by: Arun Siluvery <arun.siluvery@linux.intel.com>
Change-Id: I38288f54badfe9a7acf76d738537fae84e4c7170