DC TI GPADC: Change suspend and resume to late suspend and early init
Issue: With the TI dollarcove PMIC, while system resumed from low
power state, the temperture read operation on pmic thermistors failed.
Root cause was the adc driver was in suspend state when the system
issued a temp read.

Solution: Changed the suspend and resume calls to late suspend and
early init, such that while the temp read is issued, the driver is
out from suspend state.

Change-Id: I7711d98b1cf438b3382e09aaa0e155a9c54ca0b0
Tracked-On: https://jira01.devtools.intel.com/browse/AKP-1248
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>