FOR_UPSTREAM [VPG]: drm/i915: Fix for potential NULL pointer dereference
It is possible for memset to be called with NULL pointer and non-zero count
args which would probably cause a panic. Fix this by only initializing
the data block if the ioctl has IN or OUT data.
Found by static analysis tool.

SQUASHME! - This patch should be squashed into the following existing patch:
commit cde58b7d2e525655b126a61499bd02ceff3d8e18
Change-Id: If6b1c3e05bb8d347c8ae1e03edad43b6fa8c86a4
Author: Jon Bloomfield <jon.bloomfield@intel.com>
Date:   Thu Jul 10 14:38:16 2014 +0100

    FOR_UPSTREAM [VPG]: Introduce extended ioctl interface

    Adds a new extended ioctl mechanism, allowing several discrete
    ioctls to share a single device ioctl.

Issue: GMINL-7862
Change-Id: I37f2d669a09f69db3f9561ebe14b66c58a82a1eb
Signed-off-by: Namrta <namrta.salonie@intel.com>