SQUASHME! [VPG]: drm/i915: Fix for returning invalid fence handle on execbuff error
In the case where something fails early on during
i915_gem_do_execbuffer(), the driver would immediate exit without
calling any of the cleanup code. This is fine from a leak point of
view as nothing has been allocated yet. However, it means it is
potentially returning an invalid file descriptor for the 'fence out'
field (args->rsvd2).

This patch adds a clear of the rsvd2 field right at the start. Note
that it also first saves the incoming rsvd2 value as that is the
'fence in' parameter and must not be cleared before it has been used!

This patch could be squashed into the original patch that added the
native sync support to the execbuff code:
  commit 6b106574fa9863d9c4245b0a374fca50e09e1f2b
  Author: John Harrison <John.C.Harrison@Intel.com>
  Date:   Tue Apr 1 17:25:21 2014 +0100
  Subject: REVERTME [VPG]: drm/i915: Refactored native sync code for better file splits
  Change-Id: I8c2653787a1b15b4eb00c1dfa05b5e50ab4d2390

Change-Id: Ia7c4566604be5e95e16d0590d4d9364250fc3b50
For: VIZ-1587
For: VIZ-4741
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>