OPREGION: Fixing the error in the dptf pen function
Issue description: Previous implementation used to return bad parameter
for dptf pen when called with value 0 or 1 or any other value. Expected
result is true in case 0 or 1 and bad parameter in case of values other
than 0 or 1.

Solution: Corrected the check to validate the value appropriately and
return true in case the value is 0 or 1 and false otherwise.

Change-Id: I7a26d5d736d268963d16265d69f81f418a11c6f8
Tracked-On: https://jira01.devtools.intel.com/browse/AKP-1411
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>