REVERTME: [VPG]: drm/i915: Update pll before DPMS ON
In CHV, a DPMS ON on resume, before any mode set,
results in a blank out on HDMI, due to missing PLL
programming. This patch enables/updates chv PLLs
before enabling crtc during DPMS ON.

This patch is REVERTME because it will not be needed
once we move to our new forklifted kernel.

Change-Id: I6fa95cdc245855f1167eb4d25aefc179ebb230f7
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-25957
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>