Change the bind method for M.2 wwan_noarp driver
For the M.2 modems (enumerated with PID/VID 0x1519, 0x0443)
the IPV6 it's not functional because a MAC address is
defined for this Ethernet interface during the CDC_NCM
driver initialization. The IPV6 stack will use this MAC
address link/local instead of the link/local coming from
the network, resulting in a wrong IPV6 address.
This issue is fixed for the solder down modems (enumerated
with PID/VID 1519, 0452) by using a specific bind method
cdc_ncm_remote_mac_bind. In this method the MAC address
length is set to 1.
This patch is aligning the bind method for the M.2 modems
with the bind method for solder down modems: from
cdc_ncm_bind to cdc_ncm_remote_mac_bind.

Change-Id: If7c97d41638f6734a75b1e27992d55150ef215ae
Signed-off-by: George Sandu <georgex.sandu@intel.com>