REVERTME [VPG]: drm: Take single CRTC lock
The current DRM implementation, uses drm_modeset_lock_all()
for locking, and this is used in DPMS and flip paths. This
function takes following locks:
a. CRTC locks on all CRTCs
b. dev->mode_config.mutex
c. context_lock

But this can cause delays, blankouts and stuttering on dual
display scenarios. Consider this case:
1. DPMS off call comes on MIPI during video playback on HDMI.
2. MIPI DPMS takes lock on all CRTCs, goes for DPMS.
3. Flips are coming for HDMI playback, try to use drm_modeset_lock_all
   for all CRTCs again.
4. This thread is blocked, until MIPI display is done with DPMS off,
   which can be as big as 500-800ms. Till this time, due to locks
   flips will be failing on HDMI, causing a blank/stutter on display.

This patch removes the lock_all() from DPMS path, and flip path,
and replaces it with a single CRTC lock.

Revertme: This situation is hadled in atomic modeset, so we need
not to have this soution for long.

Change-Id: Ife6cba9fc6338966d6b8924544b52437389946c3
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-31068
Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>