Audio: HDMI: Enabling audio driver to support audio on Display Port
This patch enables audio driver to support audio on DP by programming
Maud and Naud values for DP instead of CTS and N values for HDMI.

Audio driver is informed of the external panel being DP or HDMI
through ELD sent by display driver. The second and third bits of
fifth byte of ELD should be 01 for DP and 00 for HDMI.

In case of DP connected, it queries the link rate of DP from
display driver. Naud value for DP should be 32768 according
to DP spec. Based on link rate, audio sampling frequency and
Naud value it programs the Maud values, as per the formula
specified in DP specs.

Issue: GMINL-10757
Change-Id: Idda285d71c0bdd69d23efc055d3b02a241accb91
Signed-off-by: Shubhangi Shrivastava <shubhangi.shrivastava@intel.com>
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>