lib: idr: fix out-of-bounds pointer dereference
I'm working on address sanitizer project for kernel. Recently we started
experiments with stack instrumentation, to detect out-of-bounds
read/write bugs on stack.

Just after booting I've hit out-of-bounds read on stack in idr_for_each
(and in __idr_remove_all as well):

	struct idr_layer **paa = &pa[0];

	while (id >= 0 && id <= max) {
		...
		while (n < fls(id)) {
			n += IDR_BITS;
			p = *--paa; <--- here we are reading pa[-1] value.
		}
	}
Despite the fact that after this dereference we are exiting out of loop and
never use p, such behaviour is undefined and should be avoided.

Fix this by moving pointer derference to the beggining of the loop, right
before we will use it.

Change-Id: I6afa477651e6b612d3b1d46dbed39d4c6e610a55
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-40474
Signed-off-by: Andrey Ryabinin <a.ryabinin@samsung.com>
Signed-off-by: Emmanuel Berthier <emmanuel.berthier@intel.com>