typec: toggle state machine update in ufp disconnect
due to vbus capacitance in type-a to c plug, when
removing the cable from the host side, toggle state
machine starts and detect that a cable is still present.
in order to overcome the issue, disable and re-enable
the toggle state machine and do not send notification
to the detect driver that a cable is present.

Change-Id: I4c8a9d0cc2efba7969382b36049db36981d9f56c
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-12712
Signed-off-by: R, Kannappan <r.kannappan@intel.com>