REVERTME [VPG]: drm/i915: prevent out of range px in PX macros (take 2)
In upstream we tried to fix this in the following commit:

commit fdc454c1484a20e1345cf4e4d7a9feaee814147f
Author: Michel Thierry <michel.thierry <at> intel.com>
Date:   Tue Mar 24 15:46:19 2015 +0000
    drm/i915: Prevent out of range pt in gen6_for_each_pde

but the static analyzer still complains that, just before we break due
to "iter < I915_PDES", we do "pt = (pd)->page_table[iter]" with an
iter value that is bigger than I915_PDES. Of course, this isn't really
a problem since no one uses pt outside the macro. Still, every single
new usage of the macro will create a new issue for us to mark as a
false positive.

REVERTME: Because this tree is no longer updated and is missing a lot of
changes for ppgtt. This patch is a port, and if a forklift ever happen
again, it is better to just drop this patch.

Change-Id: I69217979fcc956dc188a190f2e3b6c8f228bb730
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-40946
Signed-off-by: Michel Thierry <michel.thierry@intel.com>

## Don't forget to add a JIRA URL to the Tracked-On metadata!
## https://wiki.ith.intel.com/display/CACTUS/Patch+requirements+check
## url should be in the form:
## Tracked-On: https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
## Several Tracked-On lines are allowed
## you can fully prefill your Tracked-On commits using environment variable:
## export CURRENT_TRACKED_ON=https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
##