REVERTME: [VPG]: drm/i915: Fix PSR behavior on resume
During suspend, as part of crtc_disable sequence, PSR
is also disabled. However, if PSR was not disabled due to
some reasons (e.g. it was enabled but not active), pipe
is turned off without resetting the PSR software state
machine. Thus PSR is never enabled again on resume
because the state machine thinks it is already enabled.
This patch fixes the behavior by force resetting the
software state machine _after_ pipe is disabled.

This patch is REVERTME because CHV PSR design in upstream is
different and thus this patch will not be required when
we move to 4.0+ kernel.

Change-Id: I38c7a3c9b1ba553141a609268806b2d200a0f8ce
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-15168
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>