staging: android: sync: WA for the fence timeout issue
This is  to circumvent a  discrepancy seen sometimes during stress testing,
where the fence timeout error is reported. As per the dumps it showed that
the sync point was getting signaled round about the same time, at which
the wait on the fence got timed out. After some debugging we found that even
though the sync point got signaled before the wait on the fence got over
due to the timeout, the fence status was still seen as 0 and a fence timeout
error was reported to User. As a WA we directly check for the status of
associated sync points, once the fence timeout is detected, and in case the
sync points are already signaled, we return success to User.

Change-Id: Ie003621c38338c20cbf744df9ae645380f31e1e0
Signed-off-by: Akash Goel <akash.goel@intel.com>