Sinkpe: Changes to accommodate all the consumer mode states
Existing code doesn't handle the reset and the error recovery path.
In order to handle this, sink's own task worker is created which
will wait for the port partner's response and move the state machine
accordingly. If sink did not receive the response within the expected
time, it should reset the link and its internal state. If the port
partner is not pd capable, it will goto error recovery state and shall
not respond to port partner's further requests.

Change-Id: I046fdac87c8057044f18599843306eda331b5ad8
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-16898
Signed-off-by: R, Kannappan <r.kannappan@intel.com>