EM/GPADC: Add PM QOS constraint while reading batt temp
In Dollor Cove TC PMIC (TI PMIC), when batt temperature
conversion is requested while the device is in D3/S0i3,
it gives a wrong raw ADC value which results in reporting
a out of range temperature value and the device evntually
shuts down.
This patch adds PM QOS constraint such that while
performing ADC read, device is always is in D0 till it
completes the ADC conversion.

Change-Id: I53530fbcb756e0539b50a725b3f0d57e4eecad1e
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-3558
Signed-off-by: raosrini <srinidhi.rao@intel.com>
Reviewed-on: https://android.intel.com:443/436626