Fix for Vbus disconnect issue
When disconnecting the cable from the source side, disconnect event is
not triggered from the typec phy. Since CC voltage did not drop due
to VBUS capacitance in the cable, CC change can not be identified.
Hence Schedule/Re-Schedule task to identify valid disconnect which
helps to check the CC voltage has reached the default values, so that
the disconnect can be issued.

Change-Id: I7863958ecce1a9cfc42e0f38c95a8a726481bba0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-11921
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>
Reviewed-on: https://android.intel.com:443/448885