FOR_UPSTREAM [VPG]: drm/i915: Avoid lane optimization for External panels
Lane count optimization is not required for external panels
also it might be affecting normal working in case of MST
panels, where it is ideal to use as many lanes
available as possible to support any future panel
plugged in through that chain. This patch fixes this
by not optimizing lanes for external panels alone.

Change-Id: I12d34cdfb50e5099ac9067264cd039e6f6d678c0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-30099
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/498178