usb: dwc3: Add disable SUSPHY in DCTL and move link to U0 before send start transfer command to ep.
This patch is a fix for recovery from LPM state.

When platform has enabled LPM states, the content of MTP storage
was enumerate very slowly because USB has a problem
with recovery after LPM states.  This patch adds correct flow
consistent with the specyfication to recovery from LPM.
Also platform has a problem with certification test failed
on LPM tests on xDCI port. This patch fixes this issue
and device behaviore should be correct.

Change-Id: Iae3a895717b88ca0993668068ef5e6a6beecbdc3
Signed-off-by: Marcin Walewski <marcin.walewski@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-6490
Reviewed-on: https://android.intel.com:443/451096