SQUASHME! [VPG]: drm/i915: Avoid clearing test request till handling it
Test requests and reply need to be handled at a single point to
avoid racing issues. This patch avoids clearing of test request
in intel_dp_detect that was missed in the earlier patch.

This patch has to be squashed into
Change-Id: I8458456f305acf5bc6771ab6eec38972fc4d2b95
Author: Thulasimani,Sivakumar <sivakumar.thulasimani@intel.com>
Date:   Tue Sep 1 15:33:00 2015 +0530
FOR_UPSTREAM [VPG]: drm/i915: Cleanup panel IRQ handling

Change-Id: Ic6487c915d8439175d4ec9821c8ce76b7c061a19
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-30099
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/498193