SSIC: Disable D3 entry on every SSIC disconnect V2
This patch is a WA to disable D3 entry on every SSIC disconnect.
It was observed that the xHCI enters the D0i3 state before
the SSIC device (Intel modem) has a chance to fully enumerate,
as a consequence the enumeration fails.

The patch fixes it by blocking xhci runtime pm between disconnect
and enumeration end.

Comparing with the original patch, this patch has below improvement.
1) With the original patch, FFD still hits SWWDT related to xhci;
2) The new patch doesn't touch codes around did_runtime_put, which
looks quite like a workaround.
3) The new patch blocks runtime of bus->controller device (it's xhci)
instead of port_dev.
4) The original patch doesn't block s3. This new patch blocks s3
till modem enumeration finishes.
5) The new patch simplifies coding by removing 2 files.

Change-Id: I2ad04f61646c7391fbdd2a8640eb37dc51d83c20
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-29146
Signed-off-by: Zhong Yang <yang.zhong@intel.com>
Signed-off-by: Zhang Yanmin <yanmin.zhang@intel.com>
Reviewed-on: https://android.intel.com:443/494503