dwc3: increase autosuspend timeout
Current autosuspend timeout is too short. The autosuspend in this case is
useful to avoid a power cycle it an user is unplugging/plugging usb cable in a
short moment. But since current time is 0.5s, it's unlikely the autosuspend
will ever bring any positive side effect.

This patch increases the deadline to 10s to be more realistic.

Signed-off-by: David Cohen <david.a.cohen@intle.com>